+++ iactrace/camera/camera.py
@@ -9,6 +9,7 @@ from jax import Array
 
 from ..core.intersections import intersect_plane
 from ..core.ray_bundle import LazyRayBundle, RayBundle
+from ..core.trajectory import Trajectory
 from ..core.transforms import euler_to_matrix
 
 if TYPE_CHECKING:
@@ -89,10 +90,21 @@ def intersect_sensor(
     return sensor_rays, s_idx
 
 
+def _to_camera_frame(sensor: SensorGroup, points: Array, s_idx: Array) -> Array:
+    """Lift tile-local ``(..., n_rays, 3)`` points into the camera frame.
+
+    Each ray is placed by the tile it actually hit (``s_idx``), so a tiled or
+    curved focal surface comes out right.
+    """
+    rots = jax.vmap(euler_to_matrix)(sensor.rotations)[s_idx]  # (n_rays, 3, 3)
+    return jnp.einsum("nij,...nj->...ni", rots, points) + sensor.positions[s_idx]
+
+
 def _run_chain(
     camera: Camera,
     ray_bundle: RayBundle,
     sensor_idx: int,
+    record_trajectory: bool = False,
 ) -> tuple[Array, Array, RayBundle]:
     """Intersect the sensor, assign pixels, translate to pixel-local, run the chain.
 
@@ -104,15 +116,30 @@ def _run_chain(
     sensor tile *and* landed on the photodetector surface.
 
     Returns ``(pix_id, valid, pe_rays)``; the image / matrix folds take
-    ``pe_rays.values`` and ``collect`` reads the whole bundle.
+    ``pe_rays.values`` and ``collect`` reads the whole bundle. With
+    ``record_trajectory`` a fourth element is appended: a camera-frame
+    :class:`~iactrace.core.trajectory.Trajectory` joining the incoming leg from
+    the last optic to the path through the detection chain.
     """
     sensor = camera.sensor_groups[sensor_idx]
     sensor_rays, s_idx = intersect_sensor(camera, ray_bundle, sensor_idx)
     x, y = sensor_rays.origins[:, 0], sensor_rays.origins[:, 1]
     pix_id, valid = sensor.pixel_index_and_mask(s_idx, x, y)
     local = sensor.to_pixel_frame(sensor_rays, pix_id)
-    pe_rays = sensor.chain.propagate(local)
-    return pix_id, valid, pe_rays
+
+    if not record_trajectory:
+        return pix_id, valid, sensor.chain.propagate(local)
+
+    pe_rays, chain = sensor.chain.propagate(local, record_trajectory=True)
+    # pixel-local -> tile-local -> camera, so the chain path joins the optics.
+    points = _to_camera_frame(sensor, sensor.from_pixel_frame(chain.points, pix_id), s_idx)
+    # The bundle enters on the last optic; its first chain point is the landing
+    # on the pixel entrance, so prepending it draws the converging final leg.
+    entry = ray_bundle.origins[None]
+    # A ray that never reached a tile has no meaningful path beyond that: freeze
+    # it at the entry point rather than drawing the garbage local coordinates.
+    points = jnp.where(sensor_rays.alive[None, :, None], points, entry)
+    return pix_id, valid, pe_rays, Trajectory(points=jnp.concatenate([entry, points], axis=0))
 
 
 # Camera
@@ -132,6 +159,7 @@ class Camera(eqx.Module):
         camera.image(rb)                  # pixel image (fused, per-element fold)
         camera.response_matrix(rb)        # per-source pixel response (fused)
         camera.collect(rb)                # (pe_vals, pe_times, pix_id, detected)
+        camera.trace(rb)                  # Trajectory through the camera (diagnostics)
         rb.materialise()                  # flat camera-frame RayBundle
 
     Attributes:
@@ -178,6 +206,8 @@ class Camera(eqx.Module):
 
         Materialises a :class:`LazyRayBundle`: per-ray output cannot be
         produced incrementally.
+
+        For the *path* rays took through the camera, see :meth:`trace`.
         """
         self._require_sensor_groups(sensor_idx)
         if isinstance(ray_bundle, LazyRayBundle):
@@ -187,6 +217,24 @@ class Camera(eqx.Module):
         detected = pe_rays.alive & valid
         pe_vals = jnp.where(detected, pe_rays.values, 0.0)
         return pe_vals, pe_rays.path_length, pix_id, detected
+
+    def trace(
+        self,
+        ray_bundle: RayBundle | LazyRayBundle,
+        sensor_idx: int = 0,
+    ) -> Trajectory:
+        """The path rays take through the camera, for diagnostics / visualization.
+
+        Returns a camera-frame :class:`~iactrace.core.trajectory.Trajectory`
+        running from the last optic, through each ray's landing on its pixel, to
+        the end of the detection chain -- the final converging leg and the
+        scattering inside the concentrator as one continuous path.
+        """
+        self._require_sensor_groups(sensor_idx)
+        if isinstance(ray_bundle, LazyRayBundle):
+            ray_bundle = ray_bundle.materialise()
+
+        return _run_chain(self, ray_bundle, sensor_idx, record_trajectory=True)[3]
 
     def image(
         self,
@@ -200,6 +248,8 @@ class Camera(eqx.Module):
         by :meth:`Telescope.render`. The lazy form folds per
         primary-mirror element so the full ray buffer is never
         materialised; the eager form scatters the buffer in one call.
+
+        For the *path* rays took through the camera, see :meth:`trace`.
         """
         self._require_sensor_groups(sensor_idx)
         sensor = self.sensor_groups[sensor_idx]
@@ -267,10 +317,6 @@ class Camera(eqx.Module):
         return lazy_bundle.fold(fold_element, init)
 
     # Composition
-    #
-    # Functional updates: every setter returns a new Camera (Equinox modules are
-    # immutable pytrees), rebuilding only the affected sensor group / chain via
-    # ``eqx.tree_at``.
 
     def _with_sensor_group(self, sensor_idx: int, new_group: SensorGroup) -> Camera:
         """Swap one sensor group into a copy of the camera."""
