+++ iactrace/core/surfaces.py
@@ -6,7 +6,7 @@ import equinox as eqx
 import jax
 import jax.numpy as jnp
 
-from .intersections import intersect_conic, newton_raphson_intersect
+from .intersections import _reorigin, intersect_conic, newton_raphson_intersect
 
 # Surface group protocol
 
@@ -158,7 +158,12 @@ class SurfaceGroup(eqx.Module):
         elem = self._index(element_idx)
         t = elem._intersect_t(ray_origin, ray_direction, max_iter, tol)
         t_safe = jnp.where(jnp.isfinite(t), t, 0.0)
-        hit = ray_origin + t_safe * ray_direction
+        # Step to the hit from the vertex-adjacent point rather than from the
+        # ray origin: ``origin + t * direction`` cancels two vectors of order the
+        # source distance, so for a distant source the hit's (x, y) -- which is
+        # what the aperture test and the sag are evaluated at -- would be noise.
+        origin, t_offset = _reorigin(ray_origin, ray_direction, jnp.zeros(3))
+        hit = origin + (t_safe - t_offset) * ray_direction
         point, normal = elem.compute_sag_and_normal_at(hit[0], hit[1])
         return t, point, normal
