+++ iactrace/viz/__init__.py
@@ -1,16 +1,11 @@
-from .plotting import show_camera
-from .telescope3d import (
-    add_points,
-    add_rays,
-    add_trajectories,
-    export_mesh,
-    show_sensor_chain,
-    show_telescope,
-)
+from .overlays import add_points, add_rays, add_trajectories
+from .plotting import show_image
+from .scenes import export_mesh, show_camera, show_sensor_chain, show_telescope
 
 __all__ = [
-    "show_camera",
+    "show_image",
     "show_telescope",
+    "show_camera",
     "show_sensor_chain",
     "export_mesh",
     "add_rays",
