+++ iactrace/viz/_utils.py
@@ -1,5 +1,3 @@
-"""Shared visualization utilities."""
-
 from __future__ import annotations
